[BUGFIX]
When an error occur in the first attempt of reading file,
the OPT tab get cleaned but didn't setup a variable that tell
the program the data isn't there anymore.

Setup the variable in case of error before trying to read the file
again fixed the issue.

Not the best place to setup this var tho, should be better setting it up
after cleaning.